requirements.txt
ImageHash==4.2.1
Pillow==9.0.0
tqdm==4.62.3
